refactor: split map stack chrome, tech legend, and feed toggle glyph (#3878)
Extract minimap/players-bar/turn-feed chrome from the map stack builder,
split tech-tree legend chips and player-turn-feed newspaper glyph into
library parts, and document the new layout in repo-and-packages.md.
